+++ src/main/java/com/enigma/wmb_api/entity/BillDetail.java
@@ -27,8 +27,8 @@ public class BillDetail {
     private Menu menu;
 
     @Column(name = "qty", nullable = false)
-    private Float qty;
+    private Integer qty;
 
     @Column(name = "price", nullable = false)
-    private Float price;
+    private Integer price;
 }
